pyproject.toml
[project]
name = "video-use"
version = "0.1.0"
description = "Conversation-driven video editor skill for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "librosa",
    "matplotlib",
    "pillow",
    "numpy",
    "openai-whisper>=20250625",
]

[project.optional-dependencies]
animations = ["manim"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
